More flexibility for generic data parsing
Taxpayer Id (TIN): if the original data is of type 'double' or 'float',
convert to integer and then convert to string. This avoids the inclusion
of decimal format as part of the TIN (e.g. 123.0 becomes '123', not
'123.0').

Year: if the original data is of type 'date', convert to integer (the year extracted from the date).